+++ backends/rapidpro/msg.go
@@ -175,6 +175,7 @@ SELECT
 	attachments,
 	msg_count,
 	error_count,
+	failed_reason,
 	high_priority,
 	status,
 	visibility,
@@ -489,7 +490,6 @@ type DBMsg struct {
 	Text_                 string                 `json:"text"            db:"text"`
 	Attachments_          pq.StringArray         `json:"attachments"     db:"attachments"`
 	ExternalID_           null.String            `json:"external_id"     db:"external_id"`
-	ResponseToID_         courier.MsgID          `json:"response_to_id"  db:"response_to_id"`
 	ResponseToExternalID_ string                 `json:"response_to_external_id"`
 	IsResend_             bool                   `json:"is_resend,omitempty"`
 	Metadata_             json.RawMessage        `json:"metadata"        db:"metadata"`
@@ -498,8 +498,9 @@ type DBMsg struct {
 	ContactID_    ContactID         `json:"contact_id"      db:"contact_id"`
 	ContactURNID_ ContactURNID      `json:"contact_urn_id"  db:"contact_urn_id"`
 
-	MessageCount_ int `json:"msg_count"    db:"msg_count"`
-	ErrorCount_   int `json:"error_count"  db:"error_count"`
+	MessageCount_ int         `json:"msg_count"     db:"msg_count"`
+	ErrorCount_   int         `json:"error_count"   db:"error_count"`
+	FailedReason_ null.String `json:"failed_reason" db:"failed_reason"`
 
 	ChannelUUID_ courier.ChannelUUID `json:"channel_uuid"`
 	ContactName_ string              `json:"contact_name"`
@@ -534,7 +535,6 @@ func (m *DBMsg) ContactName() string          { return m.ContactName_ }
 func (m *DBMsg) HighPriority() bool           { return m.HighPriority_ }
 func (m *DBMsg) ReceivedOn() *time.Time       { return m.SentOn_ }
 func (m *DBMsg) SentOn() *time.Time           { return m.SentOn_ }
-func (m *DBMsg) ResponseToID() courier.MsgID  { return m.ResponseToID_ }
 func (m *DBMsg) ResponseToExternalID() string { return m.ResponseToExternalID_ }
 func (m *DBMsg) IsResend() bool               { return m.IsResend_ }